chore: refactor Backup settings to new component `Backup`
